+++ docs/victoriatraces/changelog/CHANGELOG.md
@@ -34,6 +34,8 @@ Released at 2026-04-27
 
 * SECURITY: upgrade base docker image (Alpine) from 3.23.3 to 3.23.4. See [Alpine 3.23.4 release notes](https://www.alpinelinux.org/posts/Alpine-3.20.10-3.21.7-3.22.4-3.23.4-released.html).
 
+* FEATURE: [Single-node VictoriaTraces](https://docs.victoriametrics.com/victoriatraces/) and vtselect in [VictoriaTraces cluster](https://docs.victoriametrics.com/victoriatraces/cluster/): support inverted tag filters in the Jaeger query API. Jaeger UI encodes `key!=value` as JSON `{"key!": "value"}` (and regex with `~` as in positive filters). See [#120](https://github.com/VictoriaMetrics/VictoriaTraces/issues/120).
+
 * BUGFIX: [Single-node VictoriaTraces](https://docs.victoriametrics.com/victoriatraces/) and vtinsert in [VictoriaTraces cluster](https://docs.victoriametrics.com/victoriatraces/cluster/): fix OTLP/gRPC failure with TLS enabled during HTTP/2 ALPN negotiation. See [#108](https://github.com/VictoriaMetrics/VictoriaTraces/issues/108) for details. Thank @hklhai for [the pull request #136](https://github.com/VictoriaMetrics/VictoriaTraces/pull/136).
 * BUGFIX: [Single-node VictoriaTraces](https://docs.victoriametrics.com/victoriatraces/) and vtinsert in [VictoriaTraces cluster](https://docs.victoriametrics.com/victoriatraces/cluster/): fix parsing error for OTLP/JSON when numeric values are passed as numbers instead of strings. See [#134](https://github.com/VictoriaMetrics/VictoriaTraces/issues/134) for details.